feat(hf): use to() instead of type()

+++ hf/modeling_viclip_ot.py
@@ -646,12 +646,12 @@ class ViCLIPOTModel(ViCLIPOTPretrainedModel):
     ) -> torch.Tensor:
         if isinstance(input_ids, dict):
             input_ids = {
-                k: v.type(self.text_model.dtype)
+                k: v.to(self.text_model.dtype)
                 for k, v in input_ids.items()
                 if isinstance(v, torch.Tensor)
             }
         else:
-            input_ids = input_ids.type(self.text_model.dtype)
+            input_ids = input_ids.to(self.text_model.dtype)
 
         return self.text_model(
             input_ids=input_ids, normalize=normalize, return_dict=return_dict
@@ -668,7 +668,7 @@ class ViCLIPOTModel(ViCLIPOTPretrainedModel):
         pixel_values = (
             pixel_values.pixel_values if isinstance(pixel_values, BatchFeature) else pixel_values
         )
-        pixel_values = pixel_values.type(self.vision_model.dtype)
+        pixel_values = pixel_values.to(self.vision_model.dtype)
         return self.vision_model(
             pixel_values=pixel_values, normalize=normalize, return_dict=return_dict
         ).image_features
